refactor:🌈 extension 코드 정리
